requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.2
pydantic-settings==2.6.1

# Database
motor==3.6.0
beanie==1.27.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.17
email-validator==2.2.0
fastapi-mail==1.4.1
aiofiles==24.1.0

# Background Tasks
redis==5.2.0
celery==5.4.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.27.2
mongomock==4.1.2
setuptools

